Searching and sorting algorithms
This project contains several Java classes for building a small size
application to apply the principles of sorting and searching on arrays
of different sizes. Please be patient! big arrays like this and several
in number require approximately 15' minutes to run ( CoreDuo, 3GB RAM).
At the end it will show a table with the all results. This project was
done with a team mate who code some of the classes. I coded:
BinarySearch, BubbleS20, FillArray, InsertionSort, Report, ReverseArray,
SelectionSort and the driver class: SearchingAlgorithms